[MYO-011] Refactored the brain api for better single thread use
