+++ myopic-brain/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "myopic-brain"
-version = "1.0.0"
+version = "1.1.0"
 authors = ["Thomas Ball <tomoliverball@gmail.com>"]
 edition = "2018"
 license = "MIT"
